requirements.txt
django~=4.0.5
django-timedeltafield~=0.7.10